Add clear command. Secure other mods command

+++ bot/cogs/mod_cog.py
@@ -3,8 +3,8 @@ from discord.ext.commands import Converter, BadArgument
 from bot.embeds.mod_embeds import get_temp_ban_embed
 from discord import Member, Object, NotFound
 from discord.utils import find
-from asyncio import sleep
 from typing import Optional
+from asyncio import sleep
 
 
 class BannedUser(Converter):
@@ -33,6 +33,10 @@ class Mod(Cog):
     @has_permissions(ban_members=True)
     async def temp_ban_user(self, ctx: Context, targets: Greedy[Member], duration: int,  *,
                             reason: Optional[str] = "No reason provided"):
+
+        if not targets or not duration:
+            raise BadArgument
+
         for target in targets:
             link = await ctx.channel.create_invite(max_uses=1, unique=True)
             invite_embed = get_temp_ban_embed(duration, link)
@@ -57,26 +61,62 @@ class Mod(Cog):
     @has_permissions(ban_members=True)
     async def ban_user(self, ctx: Context, targets: Greedy[Member], *,  reason: Optional[str] = "No reason provided"):
 
+        if not targets:
+            raise BadArgument
+
         for target in targets:
-            await target.ban(reason=reason)
-            # TODO remove next line
-            await ctx.send(f"Banned {target.mention}")
+            if ctx.guild.me.top_role.position > target.top_role.position \
+                    and not target.guild_permissions.administrator:
+                await target.ban(reason=reason)
+                # TODO remove next line
+                await ctx.send(f"Banned {target.mention}")
+            else:
+                await ctx.send(f"{target.mention} could not be banned")
 
     @command(name="unban")
     @bot_has_permissions(ban_members=True)
     @has_permissions(ban_members=True)
     async def unban_user(self, ctx: Context, targets: Greedy[BannedUser], *,
-                         reason : Optional[str] = "No reason provided"):
+                         reason: Optional[str] = "No reason provided"):
         if not targets:
-            await ctx.send("One or more requirement arguments are missing")
+            raise BadArgument
 
         for target in targets:
             await ctx.guild.unban(target, reason=reason)
             # TODO remove next line
             await ctx.send(f"Unbanned {target.mention}")
 
+    @command(name="kick")
+    @bot_has_permissions(kick_members=True)
+    @has_permissions(kick_members=True)
+    async def kick_user(self, ctx: Context, targets: Greedy[Member], *, reason: Optional[str] = "No reason provided"):
+
+        if not targets:
+            raise BadArgument
+
+        for target in targets:
+            if ctx.guild.me.top_role.position > target.top_role.position \
+                    and not target.guild_permissions.administrator:
+                await target.kick(reason=reason)
+                # TODO remove next line
+                await ctx.send(f"Kicked {target.mention}")
+            else:
+                await ctx.send(f"{target.mention} could not be Kicked")
+
+    @command(name="clear", aliases=['cls'])
+    @has_permissions(manage_messages=True)
+    async def clear(self, ctx, amount: int):
+
+        if not amount:
+            raise BadArgument
+
+        await ctx.channel.purge(limit=(amount + 1))
+
     @command(name="echo")
     async def echo(self, ctx: Context, *, phrase: str):
+        if not phrase:
+            raise BadArgument
+
         await ctx.message.delete()
         await ctx.send(phrase)
